Consistent output for most reify() commands
This adds a 'reify-output.js' util, which can be passed any Arborist
object after it reifies a tree.  Consistent output is printed in all
cases, showing the number of packages added/removed/changed, packages
needing funding, and a minimal (but always actionable and relevant)
audit summary.

The only code using the Installer class now is in lib/outdated.js, which
is has a pending update coming soon.

Prune and dedupe commands are awaiting top-level Arborist methods, so
that they can be similarly tightened up. (For now, this commit just has
them fail with a 'coming soon' message.)

The last piece holding the 'install/*.js' code in this repo is that it
is used in 'ls', 'fund', 'shrinkwrap', and the error-message util.