show stdout/stderr output for failed scripts
We buffer the output for scripts, and throw it away if the failure is
not something we have to care about.  But if we DO have to care about
it, it's important to show it.

This is a bare-minimum approach.  The error handling stuff here could
use a careful refactor, and it'd be nice if @npmcli/promise-spawn put
something more definitive on the error it returns, so that we didn't
have to duck-type it like this.